refactor: code and mv to node pkg
